Refactor: move CQRS to main project from shared library

+++ src/ProductService.Domain/Products/Product.cs
@@ -5,7 +5,7 @@ public class Product : BaseEntity
     public string Name { get; set; } = string.Empty;
     public string Code { get; set; } = string.Empty;
     public string? Note { get; set; }
-    public float CostPrice { get; set; }
+    public decimal CostPrice { get; set; }
     public List<string>? Images { get; set; }
 
     internal Product ChangeCode(string code)
